Support worker node taints (Review 2)
