+++ pkg/providers/vsphere/config/template-md.yaml
@@ -33,14 +33,14 @@ spec:
         nodeRegistration:
           criSocket: /var/run/containerd/containerd.sock
 {{- if .workerNodeGroupTaints }}
-          taints: {{ range .workerNodeGroupTaints}}
+          taints:{{ range .workerNodeGroupTaints}}
             - key: {{ .Key }}
               value: {{ .Value }}
               effect: {{ .Effect }}
 {{- if .TimeAdded }}
               timeAdded: {{ .TimeAdded }}
 {{- end }}
-          {{- end }}
+{{- end }}
 {{- else}}
           taints: []
 {{- end }}
