+++ pkg/validations/createvalidations/preflightvalidations.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/aws/eks-anywhere/pkg/types"
 	"github.com/aws/eks-anywhere/pkg/validations"
-	"github.com/aws/eks-anywhere/pkg/validations/upgradevalidations"
 )
 
 func (u *CreateValidations) PreflightValidations(ctx context.Context) (err error) {
@@ -20,7 +19,7 @@ func (u *CreateValidations) PreflightValidations(ctx context.Context) (err error
 		validations.ValidationResult{
 			Name:        "validate taints support",
 			Remediation: "",
-			Err:         upgradevalidations.ValidateTaintsSupport(ctx, u.Opts.Spec),
+			Err:         validations.ValidateTaintsSupport(ctx, u.Opts.Spec),
 		},
 	}
